♻️ ReportDecorator, breakout StoryLayout. ⚰️ ReportCardDecorator
