+++ packages/geoprocessing/src/components/SegmentControl.stories.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 import { SegmentControl } from "./SegmentControl";
-import ReportDecorator from "./ReportDecorator";
+import ReportDecorator from "./storybook/ReportDecorator";
 
 export default {
   component: SegmentControl,
