+++ packages/geoprocessing/src/components/Toolbar.stories.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Toolbar from "./Toolbar";
 import Card from "./Card";
 import SimpleButton from "./buttons/SimpleButton";
-import ReportDecorator from "./ReportDecorator";
+import ReportDecorator from "./storybook/ReportDecorator";
 import DataDownloadToolbar from "./DataDownloadToolbar";
 import fixtures from "../testing/fixtures";
