+++ packages/geoprocessing/src/components/storybook/CardDecorator.tsx
@@ -1,0 +1,10 @@
+import React from "react";
+import { Card } from "../Card";
+
+/**
+ * Default decorator.  Create additional building on StoryLayout for more sophisticated needs
+ */
+export const CardDecorator = (storyFn): JSX.Element => {
+  return <Card>{storyFn()}</Card>;
+};
+export default CardDecorator;
